refactor(core): split files into directories and introduce sub-plugins
Break GameCorePlugin::build() (~180 lines) into focused sub-plugins.
Each feature area now owns its own Plugin impl and system registrations.

Directory splits:
- fruit.rs → fruit/{mod,types,params}.rs
- components.rs → components/{mod,fruit,container,ui}.rs
- events.rs → events/mod.rs
- systems/input.rs → systems/input/{mod,resources,movement,drop,spawn,plugin}.rs
- systems/spawn.rs → systems/spawn/{mod,circle,fruit,plugin}.rs
- systems/preview.rs → systems/preview/{mod,setup,update}.rs
- systems/score.rs → systems/score/{mod,combo}.rs
- systems/collision.rs → systems/collision/mod.rs
- systems/merge.rs → systems/merge/mod.rs
- systems/boundary.rs → systems/boundary/mod.rs
- systems/game_over.rs → systems/game_over/mod.rs
- systems/pause.rs → systems/pause/mod.rs

New sub-plugins added:
  SpawnPlugin, CollisionPlugin, MergePlugin, ScorePlugin, EffectsPlugin,
  InputPlugin, PreviewPlugin, BoundaryPlugin, GameOverPlugin, PausePlugin

lib.rs is now a clean composition of sub-plugins with only global
state/resource init remaining in GameCorePlugin::build().

All system ordering constraints (.after() chains) are preserved inside
the individual sub-plugins. Public API unchanged — other crates compile
without modification.

Closes #140
