+++ app/core/src/fruit/mod.rs
@@ -1,0 +1,5 @@
+pub mod params;
+pub mod types;
+
+pub use params::FruitParams;
+pub use types::FruitType;
