+++ app/core/src/fruit/types.rs
@@ -6,6 +6,8 @@
 use crate::config::FruitsConfig;
 use bevy::prelude::*;
 
+use super::params::FruitParams;
+
 /// Represents the 11 fruit types in the evolution chain
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Component)]
 pub enum FruitType {
@@ -31,28 +33,6 @@ pub enum FruitType {
     Melon,
     /// Watermelon - largest fruit (stage 11), merge only
     Watermelon,
-}
-
-/// Physical and game parameters for a fruit
-#[derive(Debug, Clone, Copy)]
-pub struct FruitParams {
-    /// Collision radius in pixels
-    pub radius: f32,
-    /// Mass for physics simulation
-    pub mass: f32,
-    /// Restitution coefficient (bounciness, 0.0-1.0)
-    pub restitution: f32,
-    /// Friction coefficient (0.0-1.0)
-    pub friction: f32,
-    /// Points awarded when this fruit is created by merging
-    pub points: u32,
-    /// Visual scale multiplier (sprite diameter = radius × 2 × sprite_scale).
-    /// Does not affect the physics collider.
-    pub sprite_scale: f32,
-    /// Horizontal anchor offset passed to `Sprite::anchor` (see `FruitConfigEntry::sprite_anchor_x`).
-    pub sprite_anchor_x: f32,
-    /// Vertical anchor offset passed to `Sprite::anchor` (see `FruitConfigEntry::sprite_anchor_y`).
-    pub sprite_anchor_y: f32,
 }
 
 impl FruitType {
