+++ app/core/src/systems/merge/mod.rs
@@ -107,6 +107,23 @@ pub fn handle_fruit_merge(
     }
 }
 
+// ---------------------------------------------------------------------------
+// Plugin
+// ---------------------------------------------------------------------------
+
+use bevy::prelude::Plugin;
+
+pub struct MergePlugin;
+
+impl Plugin for MergePlugin {
+    fn build(&self, app: &mut bevy::prelude::App) {
+        app.add_systems(
+            bevy::prelude::Update,
+            handle_fruit_merge.after(crate::systems::collision::detect_fruit_contact),
+        );
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
