+++ app/core/src/systems/preview/mod.rs
@@ -3,208 +3,35 @@
 //! This module handles the display of the next fruit that will be spawned.
 //! The preview shows a smaller version of the fruit above the spawn position.
 
+pub mod setup;
+pub mod update;
+
+pub use setup::setup_fruit_preview;
+pub use update::update_fruit_preview;
+
 use bevy::prelude::*;
 
-use bevy::sprite::Anchor;
+pub struct PreviewPlugin;
 
-use crate::components::{Fruit, FruitSpawnState, NextFruitPreview};
-use crate::config::{
-    FruitsConfig, FruitsConfigHandle, GameRulesConfig, GameRulesConfigHandle, PhysicsConfig,
-    PhysicsConfigHandle,
-};
-use crate::resources::{CircleTexture, FruitSprites, NextFruitType};
-
-/// Sets up the next fruit preview display
-///
-/// Creates a preview entity showing the next fruit that will be spawned.
-/// The preview is displayed in a fixed position on the right side of the screen,
-/// similar to Tetris or Puyo Puyo next piece preview.
-///
-/// The preview starts hidden and will become visible when the first held fruit
-/// is spawned (after the first fruit lands).
-///
-/// # System Parameters
-///
-/// - `commands`: For spawning the preview entity
-/// - `next_fruit`: The type of fruit to preview
-///
-/// # Note
-///
-/// This system should run during Startup to create the initial preview entity.
-#[allow(clippy::too_many_arguments)]
-pub fn setup_fruit_preview(
-    mut commands: Commands,
-    next_fruit: Res<NextFruitType>,
-    fruits_config_handle: Res<FruitsConfigHandle>,
-    fruits_config_assets: Res<Assets<FruitsConfig>>,
-    physics_config_handle: Res<PhysicsConfigHandle>,
-    physics_config_assets: Res<Assets<PhysicsConfig>>,
-    game_rules_handle: Res<GameRulesConfigHandle>,
-    game_rules_assets: Res<Assets<GameRulesConfig>>,
-    circle_texture: Res<CircleTexture>,
-    fruit_sprites: Option<Res<FruitSprites>>,
-) {
-    // Resolve sprite image and color (real sprite or tinted placeholder).
-    let (radius, sprite_scale, anchor_x, anchor_y) =
-        if let Some(config) = fruits_config_assets.get(&fruits_config_handle.0) {
-            next_fruit
-                .get()
-                .try_parameters_from_config(config)
-                .map(|p| (p.radius, p.sprite_scale, p.sprite_anchor_x, p.sprite_anchor_y))
-                .unwrap_or_else(|| {
-                    warn!(
-                        "⚠️ No config entry for fruit {:?}, using defaults",
-                        next_fruit.get()
-                    );
-                    (20.0, 1.0, 0.0, 0.0)
-                })
-        } else {
-            warn!("Fruits config not loaded yet, using defaults for preview");
-            (20.0, 1.0, 0.0, 0.0)
-        };
-
-    let (image, color) = fruit_sprites
-        .as_deref()
-        .map(|s| s.resolve(next_fruit.get(), circle_texture.0.clone()))
-        .unwrap_or_else(|| (circle_texture.0.clone(), next_fruit.get().placeholder_color()));
-
-    // Get preview position and scale from game rules config
-    let (preview_x_offset, preview_y_offset, preview_scale) =
-        if let Some(rules) = game_rules_assets.get(&game_rules_handle.0) {
-            (
-                rules.preview_x_offset,
-                rules.preview_y_offset,
-                rules.preview_scale,
-            )
-        } else {
-            (120.0, -100.0, 1.5) // Fallback defaults
-        };
-
-    // Get container dimensions from physics config
-    let (container_width, container_height) =
-        if let Some(physics) = physics_config_assets.get(&physics_config_handle.0) {
-            (physics.container_width, physics.container_height)
-        } else {
-            (600.0, 800.0) // Fallback defaults
-        };
-
-    // Preview position: positioned relative to container
-    let preview_x = container_width / 2.0 + preview_x_offset;
-    let preview_y = container_height / 2.0 + preview_y_offset;
-
-    commands.spawn((
-        NextFruitPreview,
-        Sprite {
-            image,
-            color,
-            custom_size: Some(Vec2::splat(radius * 2.0 * sprite_scale * preview_scale)),
-            ..default()
-        },
-        Anchor(Vec2::new(anchor_x, anchor_y)),
-        Transform::from_xyz(preview_x, preview_y, 10.0),
-        Visibility::Hidden, // Start hidden, will show when held fruit spawns
-    ));
-}
-
-/// Updates the fruit preview when the next fruit type changes
-///
-/// This system monitors changes to NextFruitType and updates the preview
-/// sprite accordingly. The preview remains in a fixed position on the right side.
-///
-/// The preview visibility is controlled based on active fruit state:
-/// - When a held or falling fruit exists: Preview is visible (shows NEXT fruit)
-/// - When no active fruits exist: Preview is hidden
-///
-/// This ensures the preview stays visible during the entire drop sequence
-/// (from holding to falling to landing), and only hides when waiting for
-/// the next fruit to spawn.
-///
-/// # System Parameters
-///
-/// - `preview_query`: Query for the preview entity
-/// - `next_fruit`: The current next fruit type
-/// - `fruit_states`: Query to check fruit spawn states
-///
-/// # Behavior
-///
-/// - When NextFruitType changes: Updates color and size
-/// - When held/falling fruit exists: Shows preview
-/// - When no active fruits: Hides preview
-/// - Position remains fixed (does not follow spawn position)
-pub fn update_fruit_preview(
-    mut preview_query: Query<(&mut Sprite, &mut Visibility, &mut Anchor), With<NextFruitPreview>>,
-    next_fruit: Res<NextFruitType>,
-    fruit_states: Query<&FruitSpawnState, With<Fruit>>,
-    fruits_config_handle: Res<FruitsConfigHandle>,
-    fruits_config_assets: Res<Assets<FruitsConfig>>,
-    game_rules_handle: Res<GameRulesConfigHandle>,
-    game_rules_assets: Res<Assets<GameRulesConfig>>,
-    circle_texture: Res<CircleTexture>,
-    fruit_sprites: Option<Res<FruitSprites>>,
-) {
-    // Get the configs
-    let fruits_config = fruits_config_assets.get(&fruits_config_handle.0);
-    let game_rules = game_rules_assets.get(&game_rules_handle.0);
-    // Check if there's a held or falling fruit
-    let has_held_fruit = fruit_states
-        .iter()
-        .any(|state| *state == FruitSpawnState::Held);
-
-    let has_falling_fruit = fruit_states
-        .iter()
-        .any(|state| *state == FruitSpawnState::Falling);
-
-    for (mut sprite, mut visibility, mut anchor) in preview_query.iter_mut() {
-        // Update preview visibility based on held or falling fruit existence
-        // Keep preview visible during fruit drop (Held -> Falling transition)
-        let desired = if has_held_fruit || has_falling_fruit {
-            Visibility::Visible
-        } else {
-            Visibility::Hidden
-        };
-
-        // Only update if changed to avoid triggering change detection unnecessarily
-        if *visibility != desired {
-            *visibility = desired;
-        }
-
-        // Update preview when next fruit type or sprite resource changes.
-        // fruit_sprites.is_changed() fires when load_fruit_sprites inserts handles
-        // at Startup, catching the case where setup_fruit_preview ran first.
-        let sprites_changed = fruit_sprites.as_ref().map(|s| s.is_changed()).unwrap_or(false);
-        if next_fruit.is_changed() || sprites_changed {
-            let (image, color) = fruit_sprites
-                .as_deref()
-                .map(|s| s.resolve(next_fruit.get(), circle_texture.0.clone()))
-                .unwrap_or_else(|| {
-                    (circle_texture.0.clone(), next_fruit.get().placeholder_color())
-                });
-            sprite.image = image;
-            sprite.color = color;
-
-            if let Some(fruits_cfg) = fruits_config {
-                let preview_scale = game_rules.map(|r| r.preview_scale).unwrap_or(1.5);
-                if let Some(params) = next_fruit.get().try_parameters_from_config(fruits_cfg) {
-                    sprite.custom_size =
-                        Some(Vec2::splat(params.radius * 2.0 * params.sprite_scale * preview_scale));
-                    anchor.0 = Vec2::new(params.sprite_anchor_x, params.sprite_anchor_y);
-                } else {
-                    warn!(
-                        "⚠️ No config entry for preview fruit {:?}, keeping previous size",
-                        next_fruit.get()
-                    );
-                }
-            }
-        }
+impl Plugin for PreviewPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(
+            OnExit(crate::states::AppState::Loading),
+            setup_fruit_preview,
+        );
+        app.add_systems(Update, update_fruit_preview);
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::components::{Fruit, FruitSpawnState, NextFruitPreview};
     use crate::config::*;
-    use crate::resources::CircleTexture;
+    use crate::resources::{CircleTexture, NextFruitType};
+    use crate::systems::preview::setup::setup_fruit_preview;
+    use crate::systems::preview::update::update_fruit_preview;
     use bevy::asset::Assets;
+    use bevy::prelude::*;
 
     /// Helper to setup test app with required resources
     fn setup_test_app() -> App {
